refactor: add index.js for importing HOCs
