+++ src/hoc/index.js
@@ -1,0 +1,4 @@
+import withSortFeatures from './withSortFeatures';
+import withModal from './withModal';
+
+export { withSortFeatures, withModal };
